+++ packages/core/babel-preset-dlight/src/pluginProvider.ts
@@ -1119,21 +1119,30 @@ export class PluginProvider {
    * @param path
    * @returns
    */
-  private geneDependencyNode(path: NodePath): t.Expression {
+  private geneDependencyNode(path: NodePath): t.Node {
     let parentPath = path
     while (parentPath?.parentPath) {
       const pParentPath = parentPath.parentPath
       if (
         !(
-          this.t.isBinaryExpression(pParentPath.node) ||
-          this.t.isMemberExpression(pParentPath.node)
+          this.t.isMemberExpression(pParentPath.node) ||
+          this.t.isOptionalMemberExpression(pParentPath.node)
         )
       ) {
-        return parentPath.node as t.Expression
+        break
       }
       parentPath = pParentPath
     }
-    return path.node as t.Expression
+    const depNode = this.t.cloneNode(parentPath.node)
+    // ---- Turn memberExpression to optionalMemberExpression
+    this.traverse(this.valueWrapper(depNode as t.Expression), {
+      MemberExpression: innerPath => {
+        if (this.t.isThisExpression(innerPath.node.object)) return
+        innerPath.node.optional = true
+        innerPath.node.type = "OptionalMemberExpression" as any
+      },
+    })
+    return depNode
   }
 
   /**