+++ packages/core/transpiler/reactivity-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/reactivity-parser
 
+## 1.0.0-next.3
+
+### Patch Changes
+
+- fix: no unknown binary expression dep node and add optional member
+
 ## 1.0.0-next.2
 
 ### Patch Changes